+++ src/StefanoTree/Exception/BaseException.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace StefanoTree\Exception;
-
-class BaseException extends \Exception
-{
-}