+++ src/StefanoTree/Exception/ExceptionInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace StefanoTree\Exception;
+
+/**
+ * Base Exception interface for all exceptions thrown by the component.
+ */
+interface ExceptionInterface
+{
+}