+++ src/StefanoTree/Exception/RootNodeAlreadyExistException.php
@@ -1,7 +1,9 @@
 <?php
 
+declare(strict_types=1);
+
 namespace StefanoTree\Exception;
 
-class RootNodeAlreadyExistException extends BaseException
+class RootNodeAlreadyExistException extends \Exception implements ExceptionInterface
 {
 }