+++ src/StefanoTree/Exception/TreeIsBrokenException.php
@@ -1,12 +1,12 @@
 <?php
 
+declare(strict_types=1);
+
 namespace StefanoTree\Exception;
 
-use Exception;
-
-class TreeIsBrokenException extends BaseException
+class TreeIsBrokenException extends \Exception implements ExceptionInterface
 {
-    public function __construct($message = '', $code = 0, Exception $previous = null)
+    public function __construct($message = '', $code = 0, \Exception $previous = null)
     {
         if (!$message) {
             $message = 'Tree structure is broken. Rebuild your tree.';