+++ CHANGELOG.md
@@ -7,10 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Removed (BC Breaks)
-- removed support for PHP version 5.6 and below
+- removed support for PHP version 7.0 and below
 - removed support for [Deprecated] library Stefano-Db
 
 ### Changed (BC Breaks)
+- All library exceptions implements \StefanoTree\Exceptions\ExceptionInterface instead \StefanoTree\Exceptions\BaseException 
 - \StefanoTree\TreeInterface::updateNode($nodeId, $data) is now \StefanoTree\TreeInterface::updateNode($nodeId, array $data)
 - \StefanoTree\TreeInterface::addNodePlacementBottom($targetNodeId, $data = array()) is now \StefanoTree\TreeInterface::addNodePlacementBottom($targetNodeId, array $data = array())
 - \StefanoTree\TreeInterface::addNodePlacementBottom returns null instead false if new node was not created